- decouple index, geom, cumulative imports. - found bug in pdsparser (Issue #11) - fix issues arising from pdsTable updates - define_bodies renamed to get_bodies in columns.py
