+++ metadata_tools/templates/sky_summary.lbl
@@ -9,7 +9,7 @@ $NOTE
 #
 #  2. Copy and edit the <collection>_geometry.py script to reflect the 
 #     type of index and glob pattern for the table files.
-#  3. Copy and edit config.py to address any host-dependent processing.
+#  3. Copy and edit *config.py to address any host-dependent processing.
 #  4. Run <collection>_geometry.py from the collection top directory.
 #
 #  In the host_defs files, replace all values describing the new host.
